% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo.R
\name{batch_geo2}
\alias{batch_geo2}
\title{Geocode batches of addresses using the Census geocoding api. This is a new, streamlined version of batch_geo.}
\usage{
batch_geo2(address_dir)
}
\arguments{
\item{address_dir}{A directory containing lists of addresses. Each list should consist of no more than 1000 addresses. The
address directory should contain NOTHING BUT lists of addresses. Addresses must be formattated according to census specifications.
For more details see \url{https://www.census.gov/geo/maps-data/data/geocoder.html}. To ensure proper functioning, save
your addresses using \code{\link{save_addresses}}.}
}
\value{
a data frame with lat/long coordinates for each address that was successfully coded and NA for all others.
}
\description{
Geocode batches of addresses using the Census geocoding api. This is a new, streamlined version of batch_geo.
}
\examples{
batch_geo2("/path/to/addresses")
}
\seealso{
\code{\link{save_addresses}}
}

