% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo.R
\name{census_geo}
\alias{census_geo}
\title{Geocode batches of addresses using the Census geocoding api.}
\usage{
census_geo(file)
}
\arguments{
\item{file}{A link to a file of up to 1,000 addresses. Addresses must be formattated according to census specifications.
For more details see \url{https://www.census.gov/geo/maps-data/data/geocoder.html}. To ensure proper functioning, save
your addresses using \code{\link{save_addresses}}.}
}
\value{
a data frame with lat/long coordinates for each address that was successfully coded and NA for all others.
}
\description{
Geocode batches of addresses using the Census geocoding api.
}
\examples{
census_geo("/path/to/file.txt")
}
\seealso{
\code{\link{save_addresses}}
}

