% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_help.R
\name{convert}
\alias{convert}
\title{Takes a set of lat/long coordinates and converts them to the Chicago coordinate system.}
\usage{
convert(df, lat, long, epsg = 4326)
}
\arguments{
\item{df}{A data frame with lat/long coordinates and at least one other field.}

\item{lat}{Latitude}

\item{long}{Longitude}

\item{epsg}{the epsg of the lat/long coordinates. Default is 4326 (WSG-84)}
}
\value{
The original data frame with two additional columns: X_coordinate and Y_coordinate.
}
\description{
Takes a set of lat/long coordinates and converts them to the Chicago coordinate system.
}
\examples{
data(hom_14)
hom_14 <- dplyr::select(hom_14, -X.Coordinate, -Y.Coordinate)
hom_14 <- convert(hom_14, lat = "Latitude", long = "Longitude")

}

