% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_help.R
\name{get_regions}
\alias{get_regions}
\title{Locates spatial coordinates within user-specified regions.}
\usage{
get_regions(df, regions, lat, long, X = NULL, Y = NULL, epsg = 4326)
}
\arguments{
\item{df}{A data frame with lat/long or X/Y coordinates.}

\item{regions}{The type of region. Options include: "CAs", "districts", "tracts", "zips", and "wards."}

\item{lat}{Name of the field storing latitude. Should be in quotes.}

\item{long}{Name of the field storing longitude. Should be in quotes.}

\item{X}{The 'X coordinate' you find on some tables from the data portal.}

\item{Y}{The 'Y coordinate' you find on some tables from the data portal. Should use EITHER lat/long or X/Y but not both.}

\item{epsg}{The epsg. Only relevant if using lat/long. Default is 4326 (WSG84).}
}
\value{
A data frame. More specifically, df + one or two regional variables. If region = "CAs", the CA number and name are added. For
all other regions a single field is added: 'TRACTCE10" for tracts, "DIST_NUM" for districts, and "ZIP" for zips.
}
\description{
Locates spatial coordinates within user-specified regions.
}

