% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_chi}
\alias{map_chi}
\title{Create a basic map of Chicago with specified regional boundaries.}
\usage{
map_chi(background, lines, regions = "CAs", title = NULL, title_size = 15)
}
\arguments{
\item{background}{The background color.}

\item{lines}{The line color.}

\item{regions}{The type of region to plot. Options include: "CAs", "tracts," "districts,"
and "zips."}

\item{title}{Title.}

\item{title_size}{Title size.}
}
\value{
A map of Chicago. More specifically, an object of type "ggplot."
}
\description{
Create a basic map of Chicago with specified regional boundaries.
}
\examples{
map_chi(background = "grey", lines = "blue", title = "Chicago Community Areas")
}

