% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_addresses.R
\name{save_addresses}
\alias{save_addresses}
\title{Saves blocks of 1000 addresses in a directory (dir).}
\usage{
save_addresses(addresses, dir, root)
}
\arguments{
\item{addresses}{A vector of addresses. Addresses should be formatted according to
census specifications: "unique id, address, city, state, zip". A real example:
"123, 33 N. LaSalle St., Chicago, IL, 60602". If any pieces are missing, substitute a
blank space. For example, if state is missing, use "123, 33 N. LaSalle St., Chicago, , 60602". If
everything is missing, use "123, , , , ". For more details, see \url{https://www.census.gov/geo/maps-data/data/geocoder.html}.}

\item{dir}{An empty directory.}

\item{root}{The root name for each address block. E.G. If root = "addr", the blocks will be labelled "addr_1.txt", "addr_2.txt", etc...}
}
\value{
Nothing returned. Properly formatted address files should appear in dir.
}
\description{
Saves blocks of 1000 addresses in a directory (dir).
}
\examples{
addr_list <- c("1, 33 North LaSalle, Chicago, IL, 60602", "2, 756 West Irving Park Rd., Chicago, IL, 60613")
addr_dir <- "/path/to/empty_directory"
stem <- "addrs"

save_addresses(addresses = addr_list, dir = addr_dir, root = stem)
}

