% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{heat_map_continuous}
\alias{heat_map_continuous}
\title{Create a heat map of Chicago with continuous shading.}
\usage{
heat_map_continuous(regions, summary_df, regions_var, fill_var, legend_name,
  palette = NULL, low_color = "#fff5eb", high_color = "#7f2704",
  na_replace = NA, lines = "black", title = NULL, title_size = 15)
}
\arguments{
\item{regions}{The type of region to plot. Options include: "CAs", "tracts," "districts,"
and "zips."}

\item{summary_df}{A data frame with summary statistics by region. Should have one row for each
region.}

\item{regions_var}{The variable in df.y that identifies the region. For example, if type = "CAs"
a common summary_var would be "Community.Areas." In general, regions_var should be numeric (e.g. use
community area numbers, not names.).}

\item{fill_var}{The variable that determines color. For example, if the map shows homicides by
police district, the fill_var would be the field that contains homicide totals.}

\item{legend_name}{Self Explanatory. Should be a character vector of length one.}

\item{palette}{The color palette. Current options are: green, blue, red, organge, and purple.}

\item{low_color}{If palette is set to NULL, the low end of the color range you want to use.}

\item{high_color}{The high end of your color range}

\item{na_replace}{What value should replace NAs in the fill_var. This arises most commonly in
maps of homicides, where regions with zero homicides are sometimes missing from df.y.}

\item{lines}{Color of border lines. Default is black.}

\item{title}{Title.}

\item{title_size}{Title size.}
}
\value{
A map of Chicago. More specifically, an object of type "ggplot."
}
\description{
Create a heat map of Chicago with continuous shading.
}
\examples{
data(hom_14)

# Create summary table
hom_sum <- dplyr::summarise(group_by(hom_14, District), homicides = n())

h_map <- heat_map_continuous(type = "districts", summary_df = hom_sum, regions_var = "District",
                             fill_var = "homicides", legend_name = "Homicides",
                             na_replace = 0)
}

